+++ @xen-orchestra/defined/src/index.js
@@ -1,3 +1,5 @@
+// @flow
+
 // Usage:
 //
 // ```js
@@ -39,7 +41,7 @@ export default function defined () {
 // const getFriendName = _ => _.friends[0].name
 // const friendName = get(getFriendName, props.user)
 // ```
-export const get = (accessor, arg) => {
+export const get = (accessor: (input: ?any) => any, arg: ?any) => {
   try {
     return accessor(arg)
   } catch (error) {
@@ -58,5 +60,5 @@ export const get = (accessor, arg) => {
 //   _ => new ProxyAgent(_)
 // )
 // ```
-export const ifDef = (value, thenFn) =>
+export const ifDef = (value: ?any, thenFn: (value: any) => any) =>
   value !== undefined ? thenFn(value) : value