+++ packages/xo-web/src/xo-app/logs/log-alert-header.js
@@ -8,7 +8,7 @@ import Icon from 'icon'
 import React from 'react'
 import ReportBugButton, { CAN_REPORT_BUG } from 'report-bug-button'
 import Tooltip from 'tooltip'
-import { get } from 'xo-defined'
+import { get } from '@xen-orchestra/defined'
 import { injectState, provideState } from '@julien-f/freactal'
 import { runBackupNgJob, subscribeBackupNgLogs } from 'xo'